feat: fix bugs && small updates

+++ internal/repository/migrations/20250301131222_users.sql
@@ -4,7 +4,7 @@ CREATE EXTENSION IF NOT EXISTS pgcrypto;
 
 CREATE TABLE users (
     uuid UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    chat_id INT UNIQUE NOT NULL,
+    chat_id BIGINT UNIQUE NOT NULL,
     tg_name VARCHAR(255) UNIQUE NOT NULL,
     balance DECIMAL(18,8) DEFAULT 0,
     wallet_address VARCHAR(255) DEFAULT '',
